added mock component to hold query wrapper
